requirements.txt
torch
torchvision
matplotlib
gradio
Pillow
scikit-image
numpy
scikit-learn
fastai
accelerate
diffusers
transformers
Pillow